backend/requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-google-genai==2.0.8
openai==1.59.5
google-generativeai==0.8.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1
requests==2.32.3
httpx==0.28.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
fastapi-cors==0.0.6
pydantic==2.10.5
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.25.2
loguru==0.7.3